feat: publish mainnet
