+++ lib/protocolSdk/create/factory-addresses.ts
@@ -2,7 +2,7 @@ import { Address } from "viem";
 import { base, baseSepolia } from "viem/chains";
 
 export const FACTORY_ADDRESSES: Record<number, Address> = {
-  [base.id]: "0x4c6b9b23be9dC281C8D49FEDAed89C57a00d3b1f",
+  [base.id]: "0x540C18B7f99b3b599c6FeB99964498931c211858",
   [baseSepolia.id]: "0x6832A997D8616707C7b68721D6E9332E77da7F6C",
 } as const;
